// src/ts/tileConfig_generated.ts
export enum CuratedTileSprite {
  VOID = 0,
  NULL = 0, // Alias for VOID
  NOTHING = 0, // Alias for VOID
  FLOOR_ROCK = 1,
  FLOOR_STONE = 2,
  FLOOR = 2, // Alias for FLOOR_STONE
  GROUND = 2, // Alias for FLOOR_STONE
  FLOOR_STONE_2 = 3,
  FLOOR_STONE_3 = 4,
  FLOOR_GRASS = 5,
  FLOOR_FLOWERS = 6,
  FLOOR_LEAVES = 7,
  CHAR_1 = 24,
  CHAR_2 = 25,
  CHAR_3 = 26,
  CHAR_4 = 27,
  CHAR_5 = 28,
  CHAR_6 = 29,
  CHAR_7 = 30,
  CHAR_8 = 31,
  ARMOR_HELMET = 32,
  ARMOR_HELMET_2 = 33,
  ARMOR_HELMET_3 = 34,
  ARMOR_HELMET_4 = 35,
  ARMOR_HELMET_5 = 36,
  ARMOR_HELMET_6 = 37,
  ARMOR_HELMET_7 = 38,
  ARMOR_FEET = 39,
  ARMOR_FEET_2 = 40,
  ARMOR_HANDS = 41,
  ARMOR_HANDS_1 = 42,
  PROP_TREE = 49,
  PROP_TREE_1 = 49, // Alias for PROP_TREE
  PROP_TREE_2 = 50,
  PROP_TREE_3 = 51,
  PROP_TREES = 52,
  PROP_TREE_4 = 53,
  PROP_TREE_5 = 54,
  PROP_CACTUS = 55,
  PROP_CACTI = 56,
  CHAR_9 = 73,
  CHAR_10 = 74,
  CHAR_11 = 75,
  CHAR_12 = 76,
  CHAR_13 = 77,
  CHAR_14 = 78,
  CHAR_15 = 79,
  CHAR_16 = 80,
  ARMOR_BODY = 81,
  ARMOR_BODY_2 = 82,
  ARMOR_BODY_3 = 83,
  ARMOR_BODY_4 = 84,
  ARMOR_BODY_5 = 85,
  ARMOR_BODY_6 = 86,
  ARMOR_BODY_7 = 87,
  ARMOR_FEET_3 = 88,
  ARMOR_FEET_4 = 89,
  ARMOR_HANDS_3 = 90,
  ARMOR_HANDS_4 = 91,
  PROP_TALL_GRASS = 98,
  PROP_VINE = 99,
  PROP_VINE_2 = 100,
  PROP_TREES_2 = 101,
  PROP_TREE_6 = 102,
  FLOOR_PUDDLE = 103,
  PROP_VINE_3 = 104,
  PROP_PALM_TREE = 105,
  CHAR_17 = 122,
  CHAR_18 = 123,
  CHAR_GOBLIN = 123, // Alias for CHAR_18
  ENEMY_GOBLIN = 123, // Alias for CHAR_18
  CHAR_19 = 124,
  CHAR_20 = 125,
  CHAR_DOGMAN = 125, // Alias for CHAR_20
  ENEMY_DOGMAN = 125, // Alias for CHAR_20
  CHAR_21 = 126,
  CHAR_4EYES = 126, // Alias for CHAR_21
  ENEMY_4EYES = 126, // Alias for CHAR_21
  CHAR_22 = 127,
  CHAR_23 = 128,
  CHAR_24 = 129,
  WEAPON = 130,
  WEAPON_2 = 131,
  WEAPON_3 = 132,
  WEAPON_4 = 133,
  WEAPON_5 = 134,
  CHAR_26 = 171,
  WEAPON_6 = 179,
  WEAPON_7 = 180,
  WEAPON_8 = 181,
  WEAPON_9 = 182,
  WEAPON_10 = 183,
  FLOOR_BRIDGE_ARCHED = 202,
  FLOOR_STONE_BRIDGE = 203,
  FLOOR_BRIDGE = 203, // Alias for FLOOR_STONE_BRIDGE
  CHAR_35 = 220,
  WEAPON_11 = 228,
  WEAPON_12 = 229,
  WEAPON_13 = 230,
  WEAPON_14 = 231,
  WEAPON_15 = 232,
  FLOOR_BRIDGE_WOODEN_ARCHED = 251,
  FLOOR_WOODEN_BRIDGE = 252,
  CHAR_44 = 269,
  CHAR_LOBSTER = 269, // Alias for CHAR_44
  MONSTER_LOBSTER = 269, // Alias for CHAR_44
  CHAR_45 = 270,
  CHAR_CRAB = 270, // Alias for CHAR_45
  MONSTER_CRAB = 270, // Alias for CHAR_45
  CHAR_46 = 271,
  CHAR_WASP = 271, // Alias for CHAR_46
  MONSTER_WASP = 271, // Alias for CHAR_46
  CHAR_47 = 272,
  CHAR_TURTLE = 272, // Alias for CHAR_47
  MONSTER_TURTLE = 272, // Alias for CHAR_47
  CHAR_48 = 273,
  CHAR_SPIDER = 273, // Alias for CHAR_48
  MONSTER_SPIDER = 273, // Alias for CHAR_48
  CHAR_49 = 274,
  CHAR_SPIDER_2 = 274, // Alias for CHAR_49
  MONSTER_SPIDER_2 = 274, // Alias for CHAR_49
  CHAR_50 = 275,
  CHAR_SPIDER_3 = 275, // Alias for CHAR_50
  MONSTER_SPIDER_3 = 275, // Alias for CHAR_50
  CHAR_51 = 276,
  CHAR_FLY = 276, // Alias for CHAR_51
  MONSTER_FLY = 276, // Alias for CHAR_51
  WEAPON_16 = 277,
  WEAPON_17 = 278,
  WEAPON_18 = 279,
  WEAPON_19 = 280,
  WEAPON_20 = 281,
  STAIRS_UP = 296,
  STAIRS_DOWN = 297,
  OBJ_BOOKCASE = 299,
  CHEST_CLOSED = 302,
  CHEST_OPEN = 303,
  CHEST_EMPTY = 304,
  CHEST_OPEN_2 = 305,
  CHAR_53 = 318,
  CHAR_54 = 319,
  CHAR_55 = 320,
  CHAR_GHOST = 320, // Alias for CHAR_55
  MONSTER_GHOST = 320, // Alias for CHAR_55
  CHAR_56 = 321,
  CHAR_GHOST_2 = 321, // Alias for CHAR_56
  MONSTER_GHOST_2 = 321, // Alias for CHAR_56
  CHAR_57 = 322,
  CHAR_58 = 323,
  CHAR_SKELETON = 323, // Alias for CHAR_58
  MONSTER_SKELETON = 323, // Alias for CHAR_58
  OBJ_SIGN = 343,
  OBJ_SIGN_2 = 344,
  OBJ_SIGN_3 = 345,
  CHAR_62 = 367,
  PROP_MIRROR_TABLE = 392,
  CHAR_71 = 416,
  DOOR_LOCKED = 441,
  DOOR_UNLOCKED = 442,
  DOOR_OPEN = 443,
  DOOR_CLOSED = 444,
  DOOR_CLOSED_2 = 445,
  DOOR_CLOSED_3 = 446,
  DOOR_OPEN_2 = 447,
  DOOR_CLOSED_4 = 448,
  DOOR_CLOSED_5 = 449,
  DOOR_OPEN_3 = 450,
  DOOR_CLOSED_6 = 451,
  CHAR_80 = 465,
}